Optimised Collisions + FPS HUD
- Optimised collisions using chunked neighbours
- Added FPS HUD text
- Fixed movement being slowed due vertical camera orientation
- Cleaned up compiler warnings and added safety checks for getting collisions

+++ Engine/Source/Public/Game/GameObjects/EScreenObject.h
@@ -31,7 +31,7 @@ public:
     TWeak<ESprite> GetSprite(const EString& texturePath);
 
     TWeak<ESprite> GetSprite(const int index) {
-        if (m_sprites.size() > index) {
+        if ((int)m_sprites.size() > index) {
             return m_sprites.at(index);
         }
         return {};
